requirements.txt
Django==3.2.15
django-extensions==3.2.0
djangorestframework==3.13.1
dr-scaffold==2.1.2
python-decouple==3.6
drf-yasg==1.21.3
